LayoutTests/fast/css-grid-layout/named-grid-lines-with-named-grid-areas-get-set-expected.txt

PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column') is "b / d"
PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column-start') is "b"
PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row') is "g / h"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row-start') is "g"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row-end') is "h"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column') is "b / 4"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column-start') is "b"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column-end') is "4"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row') is "3 / h"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row-start') is "3"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row-end') is "h"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column') is "a-start / auto"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column-start') is "a-start"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row') is "c-start / auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row-start') is "c-start"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column') is "d-start / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column-start') is "d-start"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row') is "d-start / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row-start') is "d-start"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column') is "a / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column') is "d / d-start"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column-end') is "d-start"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row') is "a / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row') is "c / c-start"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row-start') is "c"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row-end') is "c-start"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column') is "d-start / d-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column-start') is "d-start"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column-end') is "d-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row') is "c-start / c-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row-start') is "c-start"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row-end') is "c-end"
PASS successfullyParsed is true

TEST COMPLETE
